animation hurts
animation controllers aren't working and animations really buggy (whenever an animation plays the player moves to a different position in the scene). I also am trying to add a animation from other models (with the same rigging) to the player. Weapon trails aren't working, but right now that isn't essential.